Add a few more functions to the api subset.
I'm considering ways to more cleanly automate the setup of these functions as
there will be a great deal of repetition in how they work. The two main
difficulties are the automation of C function pointers and the fact that there
are a number of different function signatures involved.

There are also functions that work with pointers, such as lua_tolstring. I don't
see an easy way to simulate such functions. One idea is to ignore the pointer
parameters if they're last. Another is, for outgoing values, to provide them as
second return values.