Add a few more api functions.
I realized here that lua_error will need special handling as it doesn't return,
but execution continues; and that lua_pushfstring and lua_pushvfstring allow no
easy wrapper, so I won't implement those for now. It would be nice if there were
a standardized dynamic way to build a va_list, or if C's variadic design were a
little more elegant and flexible.